Extract VideoLoaderDecoderBase, add test for missing labels
